refactor(trpc): split studio-access and recordings from rooms
Move studio join/invite/health procedures into studio-access and session
lifecycle into recordings, slim rooms to CRUD and members only, and migrate
all client tRPC call sites without compatibility aliases.
