+++ packages/trpc/src/modules/recordings/recordings.dto.ts
@@ -1,0 +1,4 @@
+import { z } from "zod";
+
+export const roomIdInputSchema = z.object({ roomId: z.string() });
+export const sessionIdInputSchema = z.object({ sessionId: z.string() });
