+++ packages/trpc/src/modules/recordings/recordings.policy.ts
@@ -1,0 +1,5 @@
+export const recordingsPolicy = {
+  canStartOrStopRecording(hasStudioAccess: boolean): boolean {
+    return hasStudioAccess;
+  },
+};
