+++ packages/trpc/src/modules/rooms/rooms.repository.ts
@@ -10,89 +10,8 @@ function generateRoomCode(): string {
 }
 
 export const roomsRepository = {
-  async findUniqueCode(code: string) {
-    return prisma.room.findUnique({ where: { code } });
-  },
-
-  async findStudioAccessByRoomCode(roomId: string, userId: string, tokenHash: string | null) {
-    const [member, participant, joinRequest, invite] = await Promise.all([
-      prisma.roomMember.findUnique({ where: { roomId_userId: { roomId, userId } } }),
-      prisma.roomParticipant.findUnique({ where: { roomId_userId: { roomId, userId } } }),
-      prisma.studioJoinRequest.findUnique({ where: { roomId_userId: { roomId, userId } } }),
-      tokenHash
-        ? prisma.roomInvite.findUnique({
-            where: { tokenHash },
-            select: {
-              id: true,
-              roomId: true,
-              revokedAt: true,
-              expiresAt: true,
-              usedCount: true,
-              maxUses: true,
-            },
-          })
-        : Promise.resolve(null),
-    ]);
-    return { member, participant, joinRequest, invite };
-  },
-
-  async upsertJoinRequest(data: { roomId: string; userId: string; status: string }) {
-    return prisma.studioJoinRequest.upsert({
-      where: { roomId_userId: { roomId: data.roomId, userId: data.userId } },
-      update: { status: data.status },
-      create: { roomId: data.roomId, userId: data.userId, status: data.status },
-    });
-  },
-
-  async listJoinRequests(roomId: string, status?: string) {
-    return prisma.studioJoinRequest.findMany({
-      where: { roomId, ...(status ? { status } : {}) },
-      orderBy: { createdAt: "asc" },
-      include: {
-        user: { select: { id: true, name: true, image: true, email: true } },
-      },
-    });
-  },
-
-  async updateJoinRequestStatus(roomId: string, userId: string, status: string) {
-    return prisma.studioJoinRequest.update({
-      where: { roomId_userId: { roomId, userId } },
-      data: { status },
-    });
-  },
-
-  async setRoomLocked(roomId: string, isLocked: boolean) {
-    return prisma.room.update({ where: { id: roomId }, data: { isLocked } });
-  },
-
   async findById(id: string) {
     return prisma.room.findUnique({ where: { id } });
-  },
-
-  async findAccessContext(roomId: string, userId: string) {
-    const [room, member, participant] = await Promise.all([
-      prisma.room.findUnique({ where: { id: roomId } }),
-      prisma.roomMember.findUnique({ where: { roomId_userId: { roomId, userId } } }),
-      prisma.roomParticipant.findUnique({ where: { roomId_userId: { roomId, userId } } }),
-    ]);
-    return { room, member, participant };
-  },
-
-  async canUserAccessSession(sessionId: string, userId: string) {
-    const session = await prisma.recordingSession.findFirst({
-      where: {
-        id: sessionId,
-        room: {
-          OR: [
-            { creatorId: userId },
-            { members: { some: { userId } } },
-            { participants: { some: { userId } } },
-          ],
-        },
-      },
-      select: { id: true },
-    });
-    return !!session;
   },
 
   async findByIdWithRelations(id: string) {
@@ -145,204 +64,6 @@ export const roomsRepository = {
       code = generateRoomCode();
     }
     return code;
-  },
-
-  // Participant methods
-  async findParticipant(roomId: string, userId: string) {
-    return prisma.roomParticipant.findUnique({
-      where: { roomId_userId: { roomId, userId } },
-    });
-  },
-
-  async addParticipant(roomId: string, userId: string) {
-    return prisma.roomParticipant.upsert({
-      where: { roomId_userId: { roomId, userId } },
-      update: {},
-      create: { roomId, userId },
-    });
-  },
-
-  async setRecordingConsent(roomId: string, userId: string, consentedAt: Date) {
-    return prisma.roomParticipant.upsert({
-      where: { roomId_userId: { roomId, userId } },
-      update: { recordingConsentedAt: consentedAt },
-      create: { roomId, userId, recordingConsentedAt: consentedAt },
-    });
-  },
-
-  async getRecordingConsent(roomId: string, userId: string) {
-    const row = await prisma.roomParticipant.findUnique({
-      where: { roomId_userId: { roomId, userId } },
-      select: { recordingConsentedAt: true },
-    });
-    return row?.recordingConsentedAt ?? null;
-  },
-
-  async createInvite(data: {
-    roomId: string;
-    tokenHash: string;
-    role: string;
-    email?: string;
-    maxUses?: number;
-    expiresAt?: Date;
-    createdBy: string;
-  }) {
-    return prisma.roomInvite.create({ data });
-  },
-
-  async listInvites(roomId: string) {
-    return prisma.roomInvite.findMany({
-      where: { roomId },
-      orderBy: { createdAt: "desc" },
-      select: {
-        id: true,
-        roomId: true,
-        role: true,
-        email: true,
-        maxUses: true,
-        usedCount: true,
-        expiresAt: true,
-        revokedAt: true,
-        createdAt: true,
-      },
-    });
-  },
-
-  async findInviteByTokenHash(tokenHash: string) {
-    return prisma.roomInvite.findUnique({
-      where: { tokenHash },
-      include: { room: true },
-    });
-  },
-
-  async revokeInvite(inviteId: string) {
-    return prisma.roomInvite.update({ where: { id: inviteId }, data: { revokedAt: new Date() } });
-  },
-
-  async consumeInvite(inviteId: string, roomId: string, userId: string) {
-    return prisma.$transaction([
-      prisma.roomInvite.update({
-        where: { id: inviteId },
-        data: { usedCount: { increment: 1 } },
-      }),
-      prisma.roomParticipant.upsert({
-        where: { roomId_userId: { roomId, userId } },
-        update: {},
-        create: { roomId, userId },
-      }),
-    ]);
-  },
-
-  async removeParticipant(roomId: string, userId: string) {
-    return prisma.roomParticipant.deleteMany({
-      where: { roomId, userId },
-    });
-  },
-
-  async listParticipants(roomId: string) {
-    return prisma.roomParticipant.findMany({
-      where: { roomId },
-      include: {
-        user: { select: { id: true, name: true, email: true, image: true } },
-      },
-      orderBy: { joinedAt: "asc" },
-    });
-  },
-
-  // Session methods
-  async endActiveSessions(roomId: string) {
-    return prisma.recordingSession.updateMany({
-      where: { roomId, status: "RECORDING" },
-      data: { status: "COMPLETED", endedAt: new Date() },
-    });
-  },
-
-  async createSession(roomId: string) {
-    return prisma.recordingSession.create({
-      data: { roomId, status: "RECORDING", startedAt: new Date() },
-    });
-  },
-
-  async findSession(sessionId: string) {
-    return prisma.recordingSession.findUnique({ where: { id: sessionId } });
-  },
-
-  async findSessionHostUserId(sessionId: string): Promise<string | null> {
-    const row = await prisma.recordingSession.findUnique({
-      where: { id: sessionId },
-      select: { room: { select: { creatorId: true } } },
-    });
-    return row?.room.creatorId ?? null;
-  },
-
-  async findActiveSessionByRoom(roomId: string) {
-    return prisma.recordingSession.findFirst({
-      where: { roomId, status: "RECORDING" },
-      orderBy: { startedAt: "desc" },
-      select: { id: true },
-    });
-  },
-
-  async markSessionComplete(sessionId: string) {
-    return prisma.recordingSession.update({
-      where: { id: sessionId },
-      data: { status: "COMPLETED", endedAt: new Date() },
-    });
-  },
-
-  async findFirstAudioTrack(sessionId: string) {
-    return prisma.recordingTrack.findFirst({
-      where: {
-        sessionId,
-        type: "MICROPHONE",
-        status: "COMPLETED",
-        OR: [{ s3Key: { not: "" } }, { s3Url: { not: null } }],
-      },
-      select: { s3Key: true, s3Url: true },
-    });
-  },
-
-  async listSessions(roomId: string, take = 50) {
-    return prisma.recordingSession.findMany({
-      where: { roomId },
-      orderBy: { startedAt: "desc" },
-      take,
-      include: {
-        tracks: {
-          include: {
-            user: { select: { id: true, name: true } },
-          },
-          orderBy: { createdAt: "asc" },
-        },
-      },
-    });
-  },
-
-  async getSessionWithDetails(sessionId: string) {
-    return prisma.recordingSession.findUnique({
-      where: { id: sessionId },
-      include: {
-        room: { select: { id: true, name: true, code: true } },
-        tracks: {
-          include: {
-            user: { select: { id: true, name: true } },
-          },
-          orderBy: { createdAt: "asc" },
-        },
-      },
-    });
-  },
-
-  async listRecentByCreator(userId: string, take = 10) {
-    return prisma.recordingSession.findMany({
-      where: { room: { creatorId: userId } },
-      orderBy: { startedAt: "desc" },
-      take,
-      include: {
-        room: { select: { id: true, name: true, code: true } },
-        tracks: { select: { id: true, status: true, type: true } },
-      },
-    });
   },
 
   // Member methods
