+++ packages/trpc/src/modules/rooms/rooms.router.ts
@@ -62,180 +62,4 @@ export const roomsRouter = {
   getRoomMembers: hostProcedure
     .input(z.object({ roomId: z.string() }))
     .query(({ input }) => roomsService.getRoomMembers(input.roomId)),
-
-  createInvite: hostProcedure
-    .input(
-      z.object({
-        roomId: z.string(),
-        role: z.enum(["participant", "editor", "viewer"]).default("participant"),
-        email: z.string().email().optional(),
-        maxUses: z.number().int().positive().max(100).optional(),
-        expiresAt: z.date().optional(),
-      }),
-    )
-    .mutation(({ input, ctx }) =>
-      roomsService.createInvite({
-        actorId: ctx.session.user.id,
-        roomId: input.roomId,
-        role: input.role,
-        email: input.email,
-        maxUses: input.maxUses,
-        expiresAt: input.expiresAt,
-      }),
-    ),
-
-  listInvites: hostProcedure
-    .input(z.object({ roomId: z.string() }))
-    .query(({ input, ctx }) =>
-      roomsService.listInvites({ actorId: ctx.session.user.id, roomId: input.roomId }),
-    ),
-
-  revokeInvite: hostProcedure
-    .input(z.object({ roomId: z.string(), inviteId: z.string() }))
-    .mutation(({ input, ctx }) =>
-      roomsService.revokeInvite({
-        actorId: ctx.session.user.id,
-        roomId: input.roomId,
-        inviteId: input.inviteId,
-      }),
-    ),
-
-  validateInvite: publicProcedure
-    .input(z.object({ code: z.string(), token: z.string() }))
-    .query(({ input }) => roomsService.validateInvite(input)),
-
-  joinRoom: protectedProcedure
-    .input(z.object({ code: z.string(), inviteToken: z.string().optional() }))
-    .mutation(({ input, ctx }) =>
-      roomsService.joinRoom({
-        userId: ctx.session.user.id,
-        code: input.code,
-        inviteToken: input.inviteToken,
-      }),
-    ),
-
-  lockRoom: hostProcedure
-    .input(z.object({ roomId: z.string() }))
-    .mutation(({ input, ctx }) =>
-      roomsService.lockRoom({ actorId: ctx.session.user.id, roomId: input.roomId }),
-    ),
-
-  unlockRoom: hostProcedure
-    .input(z.object({ roomId: z.string() }))
-    .mutation(({ input, ctx }) =>
-      roomsService.unlockRoom({ actorId: ctx.session.user.id, roomId: input.roomId }),
-    ),
-
-  listJoinRequests: hostProcedure
-    .input(
-      z.object({
-        roomId: z.string(),
-        status: z.enum(["pending", "admitted", "denied"]).optional(),
-      }),
-    )
-    .query(({ input, ctx }) =>
-      roomsService.listJoinRequests({
-        actorId: ctx.session.user.id,
-        roomId: input.roomId,
-        status: input.status,
-      }),
-    ),
-
-  admitJoinRequest: hostProcedure
-    .input(z.object({ roomId: z.string(), targetUserId: z.string() }))
-    .mutation(({ input, ctx }) =>
-      roomsService.admitJoinRequest({
-        actorId: ctx.session.user.id,
-        roomId: input.roomId,
-        targetUserId: input.targetUserId,
-      }),
-    ),
-
-  denyJoinRequest: hostProcedure
-    .input(z.object({ roomId: z.string(), targetUserId: z.string() }))
-    .mutation(({ input, ctx }) =>
-      roomsService.denyJoinRequest({
-        actorId: ctx.session.user.id,
-        roomId: input.roomId,
-        targetUserId: input.targetUserId,
-      }),
-    ),
-
-  leaveRoom: protectedProcedure
-    .input(z.object({ roomId: z.string() }))
-    .mutation(({ input, ctx }) =>
-      roomsService.leaveRoom({ userId: ctx.session.user.id, roomId: input.roomId }),
-    ),
-
-  getRoomParticipants: protectedProcedure
-    .input(z.object({ roomId: z.string() }))
-    .query(({ input }) => roomsService.getRoomParticipants(input.roomId)),
-
-  getStudioContext: protectedProcedure
-    .input(z.object({ roomId: z.string() }))
-    .query(({ input, ctx }) =>
-      roomsService.getStudioContext({ userId: ctx.session.user.id, roomId: input.roomId }),
-    ),
-
-  getStudioHealth: protectedProcedure
-    .input(z.object({ roomId: z.string() }))
-    .query(({ input, ctx }) =>
-      roomsService.getStudioHealth({ userId: ctx.session.user.id, roomId: input.roomId }),
-    ),
-
-  acknowledgeRecordingConsent: protectedProcedure
-    .input(z.object({ roomId: z.string() }))
-    .mutation(({ input, ctx }) =>
-      roomsService.acknowledgeRecordingConsent({
-        userId: ctx.session.user.id,
-        roomId: input.roomId,
-      }),
-    ),
-
-  removeGuest: hostProcedure
-    .input(z.object({ roomId: z.string(), targetUserId: z.string() }))
-    .mutation(({ input, ctx }) =>
-      roomsService.removeGuest({
-        actorId: ctx.session.user.id,
-        roomId: input.roomId,
-        targetUserId: input.targetUserId,
-      }),
-    ),
-
-  requestParticipantMute: hostProcedure
-    .input(z.object({ roomId: z.string(), targetUserId: z.string() }))
-    .mutation(({ input, ctx }) =>
-      roomsService.requestParticipantMute({
-        actorId: ctx.session.user.id,
-        roomId: input.roomId,
-        targetUserId: input.targetUserId,
-      }),
-    ),
-
-  startRecordingSession: hostProcedure
-    .input(z.object({ roomId: z.string() }))
-    .mutation(({ input, ctx }) =>
-      roomsService.startRecordingSession({ actorId: ctx.session.user.id, roomId: input.roomId }),
-    ),
-
-  stopRecordingSession: hostProcedure
-    .input(z.object({ sessionId: z.string() }))
-    .mutation(({ input, ctx }) =>
-      roomsService.stopRecordingSession({
-        actorId: ctx.session.user.id,
-        sessionId: input.sessionId,
-      }),
-    ),
-
-  getRecordingSessions: hostProcedure
-    .input(z.object({ roomId: z.string() }))
-    .query(({ input }) => roomsService.getRecordingSessions(input.roomId)),
-
-  getRecordingSessionById: hostProcedure
-    .input(z.object({ sessionId: z.string() }))
-    .query(({ input }) => roomsService.getRecordingSessionById(input.sessionId)),
-
-  listRecentSessions: hostProcedure.query(({ ctx }) =>
-    roomsService.listRecentSessions(ctx.session.user.id),
-  ),
 } satisfies TRPCRouterRecord;
