+++ packages/trpc/src/modules/rooms/rooms.service.ts
@@ -2,16 +2,12 @@ import { prisma } from "@ototabi/store";
 import { TRPCError } from "@trpc/server";
 import crypto from "node:crypto";
 
-import { scheduleTranscriptForSession } from "../../lib/schedule-transcript";
-import { recordingEventsService } from "../recording-events/recording-events.service";
-import { usageService } from "../usage/usage.service";
 import {
   defaultLobbyInviteExpiresAt,
   DEFAULT_LOBBY_MAX_USES,
-  enterStudio,
   hashInviteToken,
-  StudioAccessError,
-} from "./enter-studio";
+} from "../studio-access/enter-studio";
+import { studioAccessRepository } from "../studio-access/studio-access.repository";
 import { roomsPolicy } from "./rooms.policy";
 import { roomsRepository } from "./rooms.repository";
 
@@ -29,7 +25,7 @@ export const roomsService = {
     });
 
     const lobbyToken = createInviteToken();
-    await roomsRepository.createInvite({
+    await studioAccessRepository.createInvite({
       roomId: room.id,
       tokenHash: hashInviteToken(lobbyToken),
       role: "participant",
@@ -113,363 +109,5 @@ export const roomsService = {
 
   async getRoomMembers(roomId: string) {
     return roomsRepository.listMembers(roomId);
-  },
-
-  async createInvite(params: {
-    actorId: string;
-    roomId: string;
-    role: "participant" | "editor" | "viewer";
-    email?: string;
-    maxUses?: number;
-    expiresAt?: Date;
-  }) {
-    const room = await roomsRepository.findById(params.roomId);
-    if (!room) throw new TRPCError({ code: "NOT_FOUND", message: "Room not found" });
-
-    const actorMember = await roomsRepository.findMember(params.roomId, params.actorId);
-    if (!roomsPolicy.canInviteMember(actorMember, room, params.actorId)) {
-      throw new TRPCError({ code: "FORBIDDEN", message: "Not authorized to create invites" });
-    }
-
-    const token = createInviteToken();
-    const invite = await roomsRepository.createInvite({
-      roomId: params.roomId,
-      tokenHash: hashInviteToken(token),
-      role: params.role,
-      email: params.email,
-      maxUses: params.maxUses,
-      expiresAt: params.expiresAt,
-      createdBy: params.actorId,
-    });
-
-    return { ...invite, token };
-  },
-
-  async listInvites(params: { actorId: string; roomId: string }) {
-    const room = await roomsRepository.findById(params.roomId);
-    if (!room) throw new TRPCError({ code: "NOT_FOUND", message: "Room not found" });
-    const actorMember = await roomsRepository.findMember(params.roomId, params.actorId);
-    if (!roomsPolicy.canInviteMember(actorMember, room, params.actorId)) {
-      throw new TRPCError({ code: "FORBIDDEN", message: "Not authorized to view invites" });
-    }
-    return roomsRepository.listInvites(params.roomId);
-  },
-
-  async revokeInvite(params: { actorId: string; roomId: string; inviteId: string }) {
-    const room = await roomsRepository.findById(params.roomId);
-    if (!room) throw new TRPCError({ code: "NOT_FOUND", message: "Room not found" });
-    const actorMember = await roomsRepository.findMember(params.roomId, params.actorId);
-    if (!roomsPolicy.canInviteMember(actorMember, room, params.actorId)) {
-      throw new TRPCError({ code: "FORBIDDEN", message: "Not authorized to revoke invites" });
-    }
-    await roomsRepository.revokeInvite(params.inviteId);
-    return { success: true };
-  },
-
-  async validateInvite(params: { code: string; token: string }) {
-    const invite = await roomsRepository.findInviteByTokenHash(hashInviteToken(params.token));
-    if (!invite || invite.room.code !== params.code || !roomsPolicy.isInviteUsable(invite)) {
-      throw new TRPCError({ code: "FORBIDDEN", message: "Invite is invalid or expired" });
-    }
-    return {
-      roomId: invite.roomId,
-      roomName: invite.room.name,
-      role: invite.role,
-      email: invite.email,
-      expiresAt: invite.expiresAt,
-    };
-  },
-
-  async joinRoom(params: { userId: string; code: string; inviteToken?: string }) {
-    const hadInvite = !!params.inviteToken;
-    let roomId: string;
-    try {
-      ({ roomId } = await enterStudio({
-        userId: params.userId,
-        roomCode: params.code,
-        inviteToken: params.inviteToken,
-      }));
-    } catch (err) {
-      if (err instanceof StudioAccessError) {
-        throw new TRPCError({ code: err.code, message: err.message });
-      }
-      throw err;
-    }
-
-    const room = await roomsRepository.findUniqueCode(params.code);
-    if (!room) throw new TRPCError({ code: "NOT_FOUND", message: "Room not found" });
-
-    const activeSession = await roomsRepository.findActiveSessionByRoom(roomId);
-    if (activeSession) {
-      await recordingEventsService.createEvent({
-        actorId: params.userId,
-        sessionId: activeSession.id,
-        type: "JOIN",
-        message: hadInvite ? "Participant joined studio with invite" : "Participant joined studio",
-      });
-    }
-
-    return room;
-  },
-
-  async lockRoom(params: { actorId: string; roomId: string }) {
-    const room = await roomsRepository.findById(params.roomId);
-    if (!room) throw new TRPCError({ code: "NOT_FOUND", message: "Room not found" });
-    const actorMember = await roomsRepository.findMember(params.roomId, params.actorId);
-    if (!roomsPolicy.canToggleRoomLock(actorMember, room, params.actorId)) {
-      throw new TRPCError({ code: "FORBIDDEN", message: "Not authorized to lock this room" });
-    }
-    return roomsRepository.setRoomLocked(params.roomId, true);
-  },
-
-  async unlockRoom(params: { actorId: string; roomId: string }) {
-    const room = await roomsRepository.findById(params.roomId);
-    if (!room) throw new TRPCError({ code: "NOT_FOUND", message: "Room not found" });
-    const actorMember = await roomsRepository.findMember(params.roomId, params.actorId);
-    if (!roomsPolicy.canToggleRoomLock(actorMember, room, params.actorId)) {
-      throw new TRPCError({ code: "FORBIDDEN", message: "Not authorized to unlock this room" });
-    }
-    return roomsRepository.setRoomLocked(params.roomId, false);
-  },
-
-  async listJoinRequests(params: { actorId: string; roomId: string; status?: string }) {
-    const room = await roomsRepository.findById(params.roomId);
-    if (!room) throw new TRPCError({ code: "NOT_FOUND", message: "Room not found" });
-    const actorMember = await roomsRepository.findMember(params.roomId, params.actorId);
-    if (!roomsPolicy.canManageJoinRequests(actorMember, room, params.actorId)) {
-      throw new TRPCError({ code: "FORBIDDEN", message: "Not authorized to view join requests" });
-    }
-    return roomsRepository.listJoinRequests(params.roomId, params.status);
-  },
-
-  async admitJoinRequest(params: { actorId: string; roomId: string; targetUserId: string }) {
-    const room = await roomsRepository.findById(params.roomId);
-    if (!room) throw new TRPCError({ code: "NOT_FOUND", message: "Room not found" });
-    const actorMember = await roomsRepository.findMember(params.roomId, params.actorId);
-    if (!roomsPolicy.canManageJoinRequests(actorMember, room, params.actorId)) {
-      throw new TRPCError({ code: "FORBIDDEN", message: "Not authorized to admit guests" });
-    }
-    return roomsRepository.upsertJoinRequest({
-      roomId: params.roomId,
-      userId: params.targetUserId,
-      status: "admitted",
-    });
-  },
-
-  async denyJoinRequest(params: { actorId: string; roomId: string; targetUserId: string }) {
-    const room = await roomsRepository.findById(params.roomId);
-    if (!room) throw new TRPCError({ code: "NOT_FOUND", message: "Room not found" });
-    const actorMember = await roomsRepository.findMember(params.roomId, params.actorId);
-    if (!roomsPolicy.canManageJoinRequests(actorMember, room, params.actorId)) {
-      throw new TRPCError({ code: "FORBIDDEN", message: "Not authorized to deny guests" });
-    }
-    return roomsRepository.upsertJoinRequest({
-      roomId: params.roomId,
-      userId: params.targetUserId,
-      status: "denied",
-    });
-  },
-
-  async leaveRoom(params: { userId: string; roomId: string }) {
-    const activeSession = await roomsRepository.findActiveSessionByRoom(params.roomId);
-    await roomsRepository.removeParticipant(params.roomId, params.userId);
-    if (activeSession) {
-      await recordingEventsService.createEvent({
-        sessionId: activeSession.id,
-        type: "LEAVE",
-        message: "Participant left studio",
-        metadata: { userId: params.userId },
-      });
-    }
-    return { success: true };
-  },
-
-  async getRoomParticipants(roomId: string) {
-    return roomsRepository.listParticipants(roomId);
-  },
-
-  async getStudioContext(params: { userId: string; roomId: string }) {
-    const room = await roomsRepository.findById(params.roomId);
-    if (!room) throw new TRPCError({ code: "NOT_FOUND", message: "Room not found" });
-    const member = await roomsRepository.findMember(params.roomId, params.userId);
-    const recordingConsentedAt = await roomsRepository.getRecordingConsent(
-      params.roomId,
-      params.userId,
-    );
-    return {
-      canControlStudio: roomsPolicy.canControlStudio(member, room, params.userId),
-      memberRole: member?.role ?? (room.creatorId === params.userId ? "creator" : null),
-      recordingConsentedAt,
-      hasRecordingConsent: !!recordingConsentedAt,
-    };
-  },
-
-  async getStudioHealth(params: { userId: string; roomId: string }) {
-    const room = await roomsRepository.findById(params.roomId);
-    if (!room) throw new TRPCError({ code: "NOT_FOUND", message: "Room not found" });
-    const { member, participant } = await roomsRepository.findAccessContext(
-      params.roomId,
-      params.userId,
-    );
-    if (
-      !roomsPolicy.canJoinRoom({
-        room,
-        userId: params.userId,
-        member,
-        participant,
-        inviteUsable: false,
-      })
-    ) {
-      throw new TRPCError({ code: "FORBIDDEN", message: "Not authorized in this room" });
-    }
-
-    const rows = await roomsRepository.listParticipants(params.roomId);
-    return {
-      participants: rows.map((row) => ({
-        userId: row.userId,
-        name: row.user.name,
-        email: row.user.email,
-        hasRecordingConsent: !!row.recordingConsentedAt,
-        recordingConsentedAt: row.recordingConsentedAt,
-      })),
-    };
-  },
-
-  async acknowledgeRecordingConsent(params: { userId: string; roomId: string }) {
-    const room = await roomsRepository.findById(params.roomId);
-    if (!room) throw new TRPCError({ code: "NOT_FOUND", message: "Room not found" });
-    const { member, participant } = await roomsRepository.findAccessContext(
-      params.roomId,
-      params.userId,
-    );
-    if (
-      !roomsPolicy.canJoinRoom({
-        room,
-        userId: params.userId,
-        member,
-        participant,
-        inviteUsable: false,
-      })
-    ) {
-      throw new TRPCError({ code: "FORBIDDEN", message: "Not authorized in this room" });
-    }
-    const consentedAt = new Date();
-    await roomsRepository.setRecordingConsent(params.roomId, params.userId, consentedAt);
-    return { recordingConsentedAt: consentedAt };
-  },
-
-  async removeGuest(params: { actorId: string; roomId: string; targetUserId: string }) {
-    const room = await roomsRepository.findById(params.roomId);
-    if (!room) throw new TRPCError({ code: "NOT_FOUND", message: "Room not found" });
-    if (params.targetUserId === room.creatorId) {
-      throw new TRPCError({ code: "FORBIDDEN", message: "Cannot remove the room creator" });
-    }
-    const actorMember = await roomsRepository.findMember(params.roomId, params.actorId);
-    if (!roomsPolicy.canRemoveGuest(actorMember, room, params.actorId)) {
-      throw new TRPCError({ code: "FORBIDDEN", message: "Not authorized to remove guests" });
-    }
-    const target = await roomsRepository.findParticipant(params.roomId, params.targetUserId);
-    if (!target) throw new TRPCError({ code: "NOT_FOUND", message: "Guest not in room" });
-    await roomsRepository.removeParticipant(params.roomId, params.targetUserId);
-    const activeSession = await roomsRepository.findActiveSessionByRoom(params.roomId);
-    if (activeSession) {
-      await recordingEventsService.createEvent({
-        actorId: params.actorId,
-        sessionId: activeSession.id,
-        type: "LEAVE",
-        message: "Guest removed by host",
-        metadata: { userId: params.targetUserId, removedByHost: true },
-      });
-    }
-    return { success: true };
-  },
-
-  async requestParticipantMute(params: { actorId: string; roomId: string; targetUserId: string }) {
-    const room = await roomsRepository.findById(params.roomId);
-    if (!room) throw new TRPCError({ code: "NOT_FOUND", message: "Room not found" });
-    const actorMember = await roomsRepository.findMember(params.roomId, params.actorId);
-    if (!roomsPolicy.canRequestMute(actorMember, room, params.actorId)) {
-      throw new TRPCError({ code: "FORBIDDEN", message: "Not authorized to request mute" });
-    }
-    if (params.targetUserId === params.actorId) {
-      throw new TRPCError({
-        code: "BAD_REQUEST",
-        message: "Cannot mute yourself via host request",
-      });
-    }
-    const target = await roomsRepository.findParticipant(params.roomId, params.targetUserId);
-    if (!target) throw new TRPCError({ code: "NOT_FOUND", message: "Guest not in room" });
-    return { success: true, targetUserId: params.targetUserId };
-  },
-
-  async startRecordingSession(params: { actorId: string; roomId: string }) {
-    const { room, member, participant } = await roomsRepository.findAccessContext(
-      params.roomId,
-      params.actorId,
-    );
-    if (!room) throw new TRPCError({ code: "NOT_FOUND", message: "Room not found" });
-    if (
-      !roomsPolicy.canJoinRoom({
-        room,
-        userId: params.actorId,
-        member,
-        participant,
-        inviteUsable: false,
-      })
-    ) {
-      throw new TRPCError({ code: "FORBIDDEN", message: "Not authorized to start recording" });
-    }
-
-    await usageService.assertCanStartStudioSession(room.creatorId);
-
-    await roomsRepository.endActiveSessions(params.roomId);
-    const session = await roomsRepository.createSession(params.roomId);
-    await recordingEventsService.createEvent({
-      actorId: params.actorId,
-      sessionId: session.id,
-      type: "START",
-      message: "Recording started",
-    });
-    return session;
-  },
-
-  async stopRecordingSession(params: { actorId: string; sessionId: string }) {
-    const session = await roomsRepository.findSession(params.sessionId);
-    if (!session)
-      throw new TRPCError({ code: "NOT_FOUND", message: "Recording session not found" });
-    const canAccess = await roomsRepository.canUserAccessSession(params.sessionId, params.actorId);
-    if (!canAccess) {
-      throw new TRPCError({ code: "FORBIDDEN", message: "Not authorized to stop recording" });
-    }
-    const result = await roomsRepository.markSessionComplete(params.sessionId);
-    await recordingEventsService.createEvent({
-      actorId: params.actorId,
-      sessionId: params.sessionId,
-      type: "STOP",
-      message: "Recording stopped",
-    });
-
-    try {
-      await scheduleTranscriptForSession(params.sessionId);
-    } catch {
-      // Transcript queue is optional — don't block session completion
-    }
-
-    return result;
-  },
-
-  async getRecordingSessions(roomId: string) {
-    return roomsRepository.listSessions(roomId);
-  },
-
-  async getRecordingSessionById(sessionId: string) {
-    const session = await roomsRepository.getSessionWithDetails(sessionId);
-    if (!session)
-      throw new TRPCError({ code: "NOT_FOUND", message: "Recording session not found" });
-    return session;
-  },
-
-  async listRecentSessions(userId: string) {
-    return roomsRepository.listRecentByCreator(userId);
   },
 };
